+++ src/segmentation/map_saved_subscribe.cpp
@@ -11,7 +11,7 @@ MapSavedSubscribe::MapSavedSubscribe(ros::NodeHandle handle) : handle(handle) {
     sub_map_saved_ = handle.subscribe("/build_map_finish", 1, &MapSavedSubscribe::subscribeCallback, this);
 }
 
-void MapSavedSubscribe::subscribeCallback(const std_msgs::Int16 &msg) {
+void MapSavedSubscribe::subscribeCallback(const std_msgs::Int32 &msg) {
     MapAttribute::instance().notifySaveMap();
     ExplorationCenter::instance().repaintCoveragePath(true, true);
 }
